requirements.txt
# Minimal requirements for the exact set of files in your Archive.zip:
# - bot_finall_update.py
# - client_admin_paid.py
# - client_test_pay.py
# - send_mony.py
# - youtube_chanel.py
# - ocr.py / ocr_processor.py
# - image_processing.py / scan_image.py / pillow_comparing.py / database.py / try.py
#
# Install:
#   pip install -r requirements.min.my_runfiles.txt

# Telegram bots (uses ApplicationBuilder/ContextTypes)
python-telegram-bot>=20

# Database
psycopg2-binary

# HTTP + file watching
requests
watchdog

# Phone parsing (used in bot_finall_update.py + client_test_pay.py)
phonenumbers

# Google API client (used in client_test_pay.py + youtube_chanel.py)
google-api-python-client
cachetools

# System/process utilities (used in client_test_pay.py)
psutil

# OCR + image processing (used in image_processing.py / ocr.py)
pytesseract
Pillow
opencv-python
numpy

# NLP (used in scan_image.py)
spacy

# PDF -> images (used in ocr.py)
pdf2image
